requirements.txt
anyio==4.8.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pylast==5.4.0
sniffio==1.3.1
